+++ tests/Stub/EmptyConstructionTest.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Tests\Stub;
+
+class EmptyConstructionTest
+{
+    public function __construct()
+    {
+    }
+}