+++ tests/Stub/FooBarWithoutBuiltInTypeClass.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Tests\Stub;
+
+class FooBarWithoutBuiltInTypeClass
+{
+    public $name;
+
+    public function __construct($name)
+    {
+        $this->name = $name;
+    }
+}