+++ CHANGELOG.md
@@ -4,6 +4,24 @@ All notable changes to `container` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## 1.3.0 - 2020-06-03
+
+### Added
+- Can resolve built in type of class.  
+- If argument on construct is nullable, then if can't be resolve it will inject null class.  
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## 1.2.1 - 2020-06-03
 
 ### Added